ai-engine/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pdfplumber==0.7.6
python-docx==0.8.11
sentence-transformers==2.2.2
scikit-learn==1.3.2
spacy==3.7.4
pydantic==1.10.12
numpy==1.26.4
scipy==1.11.3
gunicorn==20.1.0
